move queue, update packages and pass project data to webserver
