+++ datarithms/README.md
@@ -1,0 +1,3 @@
+# Datarithms
+
+This package includes helpful algorithms and data structures for our [Mirror](https://github.com/COSI-Lab/Mirror) project
