+++ webserver.go
@@ -4,11 +4,15 @@ import (
 	"html/template"
 	"log"
 	"net/http"
+	"sync"
 
 	"github.com/gorilla/mux"
 )
 
 var tmpls *template.Template
+var distributions []Project
+var software []Project
+var dataLock = &sync.RWMutex{}
 
 func handleIndex(w http.ResponseWriter, r *http.Request) {
 	tmpls.ExecuteTemplate(w, "index.gohtml", "")
@@ -19,7 +23,9 @@ func handleMap(w http.ResponseWriter, r *http.Request) {
 }
 
 func handleDistributions(w http.ResponseWriter, r *http.Request) {
-	tmpls.ExecuteTemplate(w, "distributions.gohtml", "chris")
+	dataLock.RLock()
+	tmpls.ExecuteTemplate(w, "distributions.gohtml", distributions)
+	dataLock.RUnlock()
 }
 
 func handleHistory(w http.ResponseWriter, r *http.Request) {
@@ -27,7 +33,9 @@ func handleHistory(w http.ResponseWriter, r *http.Request) {
 }
 
 func handleSoftware(w http.ResponseWriter, r *http.Request) {
-	tmpls.ExecuteTemplate(w, "software.gohtml", "")
+	dataLock.RLock()
+	tmpls.ExecuteTemplate(w, "software.gohtml", software)
+	dataLock.RUnlock()
 }
 
 func handleStatistics(w http.ResponseWriter, r *http.Request) {
@@ -55,6 +63,22 @@ func loggingMiddleware(next http.Handler) http.Handler {
 		log.Println("[INFO]", r.Method, r.RequestURI)
 		next.ServeHTTP(w, r)
 	})
+}
+
+// Setup distributions and software arrays
+func webserverLoadConfig(config ConfigFile) {
+	dataLock.Lock()
+	distributions = make([]Project, len(config.Mirrors))
+	software = make([]Project, len(config.Mirrors))
+
+	for _, project := range config.Mirrors {
+		if project.IsDistro {
+			distributions = append(distributions, project)
+		} else {
+			software = append(software, project)
+		}
+	}
+	dataLock.Unlock()
 }
 
 func HandleWebserver(entries chan *LogEntry) {
